Fix an assert failure with a funclet in a swifttailcc function.
The failure happens in the livedebugvalues pass.

swiftcc is already covered by isCallingConvWin64. I believe that the assert is checking that the particular calling convention uses a frame pointer, which swifttailcc does, and being relatively a recent addition, it wasn't covered by this code.

This cherrypicks the main PR https://github.com/apple/llvm-project/pull/8211