[clang] Move CCC_OVERRIDE_OPTIONS implementation to Driver (#85425)
Move CCC_OVERRIDE_OPTIONS support to clangDriver so that it may be used outside of the
clang driver binary.

The override functionality will be used in LLDB, to apply adjustments to ClangImporter
flags. This will be useful as an escape hatch when there are issues that can be fixed
by adding or removing clang flags.

The only thing changed is the name, from `ApplyQAOverride` to `applyOverrideOptions`.

(cherry picked from commit 8d7ee4691dadd3a9d831108f77d1f4e511191a44)