[lldb] Be more precise when copying over Swift resources
To run the LLDB test suite, we need to be able to find some of the Swift
compiler resources. When part of a toolchain, LLDB knows where to find
those. For at-desk builds, it copies over the resources from the Swift
build directory into LLDB.framework.

Right now we copy over the whole lib/swift directory from the Swift
build root. This contains much more than we need, such as dylibs and
dSYMs. This patch modifies LLDB to be more precise in what it copies
over.

rdar://124975526