[lldb][progress][NFC] Clarify Doxygen comments for `details` field (#86002)
The Doxygen comments for the `details` field of a progress report
currently does not specify that this field will act as the initial set
of details for a progress report that gets updated with
`Progress::Increment()`. This commit clarifies this.

(cherry picked from commit 6317c780d81327bd06701a6aa374fc92aa3d73ad)