[lldb] Add a setting to allow Swift AST context instantiations
For testing purposes, it's useful to have a setting to disable the
instantiation of SwiftASTContext, to ensure that
TypeSystemSwiftTypeRef, which by default falls back to it on failure, is
working independently.

This setting might also be useful for end users, as when debugging
a program that was built with a different compiler version versus what
is embedded in LLDB instantiation of Swift AST contexts would be moot.