[lldb] Split TestSwiftEmbeddedFrameVariable into two tests
Split TestSwiftEmbeddedFrameVariable into one test that runs with
instantiation of SwiftASTContext enabled, and one that runs with it
disabled.

The first one is useful because of the validation code that compares
TypeSystemSwiftTypeRef results with SwiftASTContext's.

The second one is useful to validate that the test is passing without
TypeSystemSwiftTypeRef falling back to SwiftASTContext.