Reland "Cherry-pick 51d5d7bbae92493a5bfa7cc6b519de8a5bb32fdb from LLVM mainline" (#8199)
* Update coroutine intrinsics documentation and few remaining tests to opaque pointers (#65698)

* Extend `retcon.once` coroutines lowering to optionally produce a normal result (#66333)

One of the main user of these kind of coroutines is swift. There yield-once (`retcon.once`) coroutines are used to temporary "expose" pointers to internal fields of various objects creating borrow scopes.

However, in some cases it might be useful also to allow these coroutines to produce a normal result, but there is no convenient way to represent this (as compared to switched-resume kind of coroutines where C++ `co_return` is transformed to a member / callback call on promise object).

The extension is simple: we allow continuation function to have a non-void result and accept optional extra arguments via a special `llvm.coro.end.result` intrinsic that would essentially forward them as normal results.