[clang][cas] Don't include PGO-related files in the include-tree when they are not used
PGO options are ignored for PCH invocations and may not be treated as dependencies from the build
system, so don't add them in the include-tree. Otherwise we may get in a situation where such a file
gets updated and an incremental build fails because of mismatched files in the include-tree of the
PCH and the translation unit that uses it.

rdar://125629656
(cherry picked from commit 4ccf152a3fd2306b7edf8a75f297047dc5168d08)