[cas] Add cc1 option -finclude-tree-preserve-pch-path
Add the PCH path to the include tree, and add a way to control whether
we use the original PCH path, or if we canonicalize it away to <PCH>.
For now, keep the default being to canonicalize it to <PCH> as this
improves canonicalization. However, a client (e.g. Swift) that knows its
PCH path will be stable can use this option to allow enabling -gmodules,
which requires a real path.

rdar://126370706
(cherry picked from commit 6f52d300218cdf6bec562be2dacfb47dd11cd8a2)