[lldb] Replace assertEquals with assertEqual (NFC)
assertEquals is a deprecated alias for assertEqual and has been removed
in Python 3.12. This wasn't an issue previously because we used a
vendored version of the unittest module. Now that we use the built-in
version this gets updated together with the Python version used to run
the test suite.

This is the equivalent of 80fcecb13c388ff087a27a4b0e7ca3dd8c98eaa4
upstream but for the Swift tests.