Add zlib compression to the debug_str_offset section
In MCCAS, the DWARF5 size is larger than the DWARF4 size because of the
presence of the debug_str_offset section. There is not much we can do to
improve the deduplication rate of that section, so the only thing to do
is to compress the data in that cas block, which is what this patch adds
support for.

(cherry picked from commit de4445559932d28698aad8836ffea0a20ff3f2ba)