[llvm][cas] Fix thread safety issues with MappedFileRegionBumpPtr
llvm::sys::fs::lockFile and friends, which are implemented with fcntl
F_SETLK, are only safe to use across processes and not across threads.
We were attempting to workaround that using createShared, which tries to
unique the mapped file per process, but (a) that only works if you have
a single copy of llvm in your process, which may not always be true, and
(b) we had a race in the implementation between closing one CAS and
opening a new one at the same path.  Replace these with an flock-based
implementation. Since this is not available on all platforms LLVM
supports we use a separate implementation for the on-disk CAS rather
than change llvm::sys::fs::lockFile.

rdar://126882843
(cherry picked from commit 15a76e048eb9d7bf62fc36865cd749b161dc5b86)