[lldb][progress] Mitigate non-specific LLDB progress reports in Swift
When LLDB reports progress on importing Swift modules, it was delivering
non-specific progress reports with only the title of "Importing Swift
modules" and no details. To report progress on activity within Swift,
LLDB first creates a progress report then updates that progress report
using a callback that LLDB sets and the Swift compiler invokes when
performing a full import.

When LLDB triggers Swift to import modules that were already imported before,
Swift will not perform a full import. Since the progress report would've
already been displayed, but the callback is never invoked, this leads
to the non-specific messages that were being displayed when importing
Swift modules.

This commit sets a unique pointer to create a progress report from
within the callback function instead of creating the report before
setting the callback. It also clears the callback on scope exit instead
of setting it to a new progress report.