Remove incorrect and superfluous caching in Mangled.
There was a global cache for Swift display names, however, it didn't
take the symbol context into account, which means the demangling
result is non-deterministic depending on whether the symbol context
had the necessary debug info to resolve the names of archetypes or
not.

In addition to that Display names are not supposed to be on a hot path
(unlike names, which are used to resolve breakpoints).