[lldb] Reinstate lldb-sbapi-dwarf-enums target (NFC) (#91390)
Alex pointed out in #91254 that we only need the custom target if we had
more than one target depending on it. This isn't the case upstream, but
on our downstream fork, we have a second dependency. Reintroduce the
target so that everything can depend on that, without the
single-dependency foot-gun.

(cherry picked from commit dad11097096c05564758e539f9f03ef883365fdd)