Set Swift clang-linker path
The swift driver is picking up the system clang instead of using the
clang set in `CC`. The system clang is attempting to use bfd to link
Swift, which doesn't work due to Swift's protected symbol usage.
Set the swift-driver tool directory based on the location of the clang
in use to ensure that we're picking up a consistent set of tools across
all of the tests.