Drop DIAssign for outlined cold function.
Fix for rdar://128178601
(Swift 6.0: Debian 12: Clang --  inst not in same function as dbg.assign)

The bug occurs because we need to fix up the DIAssign for an instruction
that has been outlined due to hot cold splitting. That change exists in
llvm trunk, but would require cherry-picking over 100 commits, so a
compromise for this release is to drop the debug info instead.