[target] .swift_ast section is a metadata section
This change ensures that this section is not marked as loadable, which makes it possible for it to be stripped. The swift_ast section holds the raw Swift module bytes that are needed by LLDB, but that are not required at runtime.

(cherry picked from commit 555eaef3ab34108b68631411f0ec8464ee45af0a)