[lldb][lldb-dap] Disable read memory test on Windows
This isn't strictly a Windows issue but for now it's the only bot
that was hit by this failure.

It can happen on Linux too but I expect we'll fix it and remove the
skip soon anyway.

Test was added in #104317.

(cherry picked from commit ab38ec9ac312460e4f71c8ad3f50b9b5723469f6)