[BoundsSafety] Apply started_by on PointerType before qualifiers so that getUnqualifiedType doesn't strip started_by
All other bounds annotations are applied directly to bare
PointerType via recursive TypeVisitor but started_by. This caused
started_by to be stripped wheneter there is a type query to get
unqualified type and the pointer type has a qualifier. This bug
caused the started_by pointer to be treated as a single pointer,
leading to a run-time trap when it gets the upper bound which has
zero memory available..

The fix is to apply started_by at the bare PointerType just like
any other bounds annotated types. Also, this strips PtrAutoAttr
added to started_by because it's no longer an implicit single
pointer.

rdar://137701968