[CAS] Fix working-directory canonicalization when using clang cc1depscan
This is a followup to bb91b1ce3bb2a662890efb67395125b0f1aac21c to fix
the removal of working-directory when clang cc1depscan is used. Since
the scanner will removes the working-directory, clang-driver needs to
make sure the output is correct if that is a relative path to the
`-working-directory` option.

rdar://136203252