[lldb] Mark forward C++ interop test XFAIL
After Swift PR #77323, we import C++ source locations for imported code.
This improves some diagnostics but also breaks some of the heuristics
LLDB have to select the right frame on traps. This PR marks the test
XFAIL until we figure out a better heuristic to stop at the right frame.
This is not a trivial question as we cannot just mark all C++ imported
source locations artifical, some of that code is user written, some is
not (coming from frameworks, STL). Moreover, some users might prefer to
stop in the C++ code and some users might prefer to stop in Swift code.